+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
-- Add metrics for ResourceQuota
+
+### Added
+
+* Add metrics for ResourceQuota
 
 ## v3.45.3 - 2025-09-29